+++ src/main/java/edu/umgc/librarymanager/IUserService.java
@@ -4,7 +4,7 @@
  * Date Created: 11/17/2020
  */
 
-package edu.umgc.librarymanager.data.access;
+package edu.umgc.librarymanager;
 
 import edu.umgc.librarymanager.data.model.user.IUser;
